sv_vcatpvfn_flags(): no "redundant arg" warnings for '%*2$d'
This addresses RT#125469: if the format string uses an explicit index for a
field width, that should suppress any "redundant argument" warnings, in
just the same way as an explicit index does when used to specify a format
argument.

Blead has a more careful reworking of the way printf warnings work, but that
change is unsuitable for maint, as it can lead to additional warnings being
emitted for some invalid format strings. This smaller change ensures that
maint (5.22.1-to-be) always emits strictly fewer warnings than 5.22.0.